requirements/base.txt
# Bleeding edge Django
django==1.7

# Configuration
django-configurations==0.8
django-secure==1.0
django-cache-url==0.8.0
dj-database-url==0.3.0

# Forms
django-braces==1.4.0
django-crispy-forms==1.4.0
django-floppyforms==1.2.0

# Models
django-model-utils==2.2

# images
Pillow==2.6.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.18.0

# For the persistance stores
psycopg2==2.5.4

# Unicode slugification
unicode-slugify==0.1.1
django-autoslug==1.7.2

# Useful things
django-avatar==2.0

# Time zones support
pytz==2014.7

# Markdown support
Markdown==2.5.1
django-markdown==0.8.1
django-markupfield==1.2.1

# Tagging
topia.termextract==1.1.0

# for API-goodness
django-filter==0.9.1
djangorestframework==3.0.0

# Thumbnails for use in previewing datapoints
sorl-thumbnail==11.12.1b

# For datapoint upload processing
newspaper==0.0.8
nltk==3.0.0
requests==2.5.0
beautifulsoup4==4.3.2
lxml==3.4.1
selenium==2.44.0
textract==1.1.0
python-magic==0.4.6

# Stuff that was required with a fresh Vagrant instance
docutils==0.12
sqlparse==0.1.13

# Commenting system (Referring directly to a git repo due to lack of latest version on PyPi)
-e git+https://github.com/django/django-contrib-comments#egg=django-contrib-comments

# Search system
elasticsearch==1.3.0
-e git+https://github.com/allyjweir/django-haystack#egg=django-haystack
-e git+https://github.com/allyjweir/saved_searches#egg=saved_searches
simplejson==3.6.5
jsonpickle==0.9.0

# Annotation extension requirements
jsonfield==1.0.2
